Add user-agent header requirement
https://github.com/forem/forem/pull/15817 disallowed empty user
agents (the intent was to do this for the site, not for the api).

This restriction only applies to fastly configured forem instances (at
present, this is probably only dev.to) - but there's no reason to
overspecify.

This sentence can be removed if anonymous api requests are permitted
in the future.